rename comprehensive flows to scenarios
